+++ src/core/runtime.js
@@ -104,6 +104,8 @@ export function ensureMap(c, r){
       layers[i].vary = growBuf(layers[i].vary || new Uint8Array(ow * oh));
       if (layers[i].deco) layers[i].deco = growBuf(layers[i].deco);
       else layers[i].deco = new Uint8Array(nw * nh);
+      if (layers[i].tint) layers[i].tint = growBuf(layers[i].tint);
+      else layers[i].tint = new Uint8Array(nw * nh);
       if (layers[i].cover) layers[i].cover = growBuf(layers[i].cover);
       if (layers[i].coverVary) layers[i].coverVary = growBuf(layers[i].coverVary);
       layers[i].roomOf = null;